# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.CreateInnerUserRequest:
    properties:
      email:
        type: string
      job:
        type: string
      name:
        type: string
      namespace:
        type: string
      operator:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      user:
        type: string
    required:
    - name
    - namespace
    - password
    - user
    type: object
  dto.UpdateInnerUserRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      job:
        type: string
      name:
        type: string
      namespace:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      user:
        type: string
    required:
    - id
    - namespace
    type: object
  dto.UserLoginRequest:
    properties:
      namespace:
        example: acl
        type: string
      password:
        example: "12345678"
        type: string
      user:
        example: admin
        type: string
    required:
    - namespace
    - password
    - user
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      ret_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Code Review API documentation
  title: Code Review API
  version: "1.0"
paths:
...
  /v1/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录信息
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回 JWT token 和用户信息
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 用户管理
  /v1/user:
    delete:
      consumes:
      - application/json
      description: 删除指定用户
      parameters:
      - description: JWT认证Token
        in: header
        name: jwt_token
        required: true
        type: string
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 用户名
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新的内部用户
      parameters:
      - description: JWT认证Token
        in: header
        name: jwt_token
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInnerUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建内部用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新内部用户信息
      parameters:
      - description: JWT认证Token
        in: header
        name: jwt_token
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInnerUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新内部用户
      tags:
      - 用户管理
  /v1/user/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      parameters:
      - description: JWT认证Token
        in: header
        name: jwt_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户信息和角色列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 用户管理
...
swagger: "2.0"